pyproject.toml
[project]
name = "betting-odds-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "konradhorber",email = "konrad.horber@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "torch (>=2.8.0,<3.0.0)",
    "google-cloud-storage (>=3.3.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "betting_odds_backend", from = "src"}]


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
